// ArchiCore/src/app/api-client/model/auditLog.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuditLog { 
    id?: number;
    action?: AuditLog.ActionEnum;
    performedBy?: string;
    userId?: number;
    performedAt?: string;
    entityType?: string;
    entityId?: string;
    details?: string;
    ipAddress?: string;
}
export namespace AuditLog {
    export const ActionEnum = {
        CreerDossier: 'CREER_DOSSIER',
        CreerMission: 'CREER_MISSION',
        ConstituerEquipe: 'CONSTITUER_EQUIPE',
        ConsulterDocument: 'CONSULTER_DOCUMENT',
        TelechargerDocument: 'TELECHARGER_DOCUMENT',
        SupprimerDocument: 'SUPPRIMER_DOCUMENT',
        SupprimerDossier: 'SUPPRIMER_DOSSIER',
        UpdateDossier: 'UPDATE_DOSSIER',
        UploadDocument: 'UPLOAD_DOCUMENT',
        SupprimerMission: 'SUPPRIMER_MISSION',
        UpdateMission: 'UPDATE_MISSION'
    } as const;
    export type ActionEnum = typeof ActionEnum[keyof typeof ActionEnum];
}